[wedding] add engagement photo to home page
