+++ wedding/src/components/Logo/Logo.module.css
@@ -1,9 +1,8 @@
 .logo {
   display: flex;
   align-items: center;
-  height: 60px;
   cursor: pointer;
-  font-size: 1.5rem;
+  font-size: 2.5rem;
   color: var(--foreground-secondary);
 }
