+++ wedding/src/components/Navigation/MobileNav.module.css
@@ -1,4 +1,4 @@
-@media (min-width: 550px) {
+@media (min-width: 924px) {
   .navbar {
     display: none;
   }
@@ -8,7 +8,7 @@
   background: none;
   border: none;
   position: absolute;
-  top: 12px;
+  top: 18px;
   left: 20px;
   z-index: 3;
   padding: 0;
