+++ wedding/src/pages/hotels.module.scss
@@ -39,7 +39,7 @@
   margin-block: 16px;
 }
 
-@media (min-width: 700px) {
+@media (min-width: 924px) {
   .travelContainer .hotelList {
     flex-direction: row;
   }
