+++ wedding/src/pages/registry.module.scss
@@ -27,7 +27,7 @@
   }
 }
 
-@media (min-width: 700px) {
+@media (min-width: 924px) {
   .registryContainer .registryGrid {
     flex-direction: row;
   }
